requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
cryptography==41.0.4
python-jose==3.3.0
gunicorn==21.2.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0 